add internal argument handling in shell + fix Shell_kill_process

+++ src/shell/shell.c
@@ -47,7 +47,37 @@ void Shell_main_loop(Shell * self) {
 }
 
 void Shell_configure(Shell * self, int argc, char * argv) {
-  rl_bind_key('\t', rl_complete);
+  int result;
+  
+  ArgsHandler *args_handler = ArgsHandler_init(6);
+  /* print help and close the program */
+  ArgsHandler_insert_arguments(args_handler, Arg_init("-h", "--help", STRING(NULL), 
+                               args_handler, 0, ArgsHandler_print)); 
+  /* set the Logger output file */
+  ArgsHandler_insert_arguments(args_handler, Arg_init("-o", "--out-file", STRING(NULL), 
+                               self->config, 1, ShellConfig_set_out_file)); 
+  /* set the Logger error file */
+  ArgsHandler_insert_arguments(args_handler, Arg_init("-e", "--err-file", STRING(NULL), 
+                               self->config, 1, ShellConfig_set_err_file)); 
+  /* set if the Logger have to print the return code of the processes*/
+  ArgsHandler_insert_arguments(args_handler, Arg_init("-c", "--code", STRING(NULL), 
+                               self->config, 0, ShellConfig_enable_print_code)); 
+  /* set the max size in characters of the output of every process */
+  ArgsHandler_insert_arguments(args_handler, Arg_init("-m", "--max-size", STRING(NULL), 
+                               self->config, 1, ShellConfig_set_output_max_size)); 
+  /* set the logger process executable path 
+  */
+  ArgsHandler_insert_arguments(args_handler, Arg_init("-l", "--logger-path", STRING(NULL), 
+                               self->config, 1, ShellConfig_set_logger_path)); 
+
+  result = ArgsHandler_handle_input(args_handler, argc, argv);
+  
+  ArgsHandler_destroy(args_handler);
+  
+  if (result == -1) {
+    Shell_print_error(self, "Invalid argument passing...", 1); 
+  }
+
  // not yet implemented 
 }
 void Shell_cd(Shell * self, char * new_dir) {
@@ -69,7 +99,8 @@ void Shell_print_help(Shell * self) {
   printf("print command has not been implemented yet.\n");
 }
 void Shell_kill_running_process(Shell * self) {
-  kill(self->running_process_pids, SIGKILL);
+  if (self->running_process_pids != NO_PROCESS)
+    kill(self->running_process_pids, SIGKILL);
 }
 void Shell_exit(Shell * self, int code) {
   Shell_kill_running_process(self);
